refactor: services refactor + improve controllers

+++ src/main/java/com/bmstu/lab/category/model/dto/CategoryDTO.java
@@ -18,7 +18,7 @@ public class CategoryDTO {
 
   private String title;
 
-  private int basePrice;
+  private double basePrice;
 
   private String imageId;
